`vec::merge`: set waker on each call
